requirements.txt
flask
google-generativeai
pdfplumber
werkzeug
gunicorn
python-dotenv
pytesseract
pillow
re